pyproject.toml
[tool.poetry]
name = "python-template-repo"
version = "0.1.0"
description = ""
authors = ["Maximilian Jugl <Maximilian.Jugl@medizin.uni-leipzig.de>"]
readme = "README.md"
packages = [{ include = "project" }]

[tool.poetry.dependencies]
python = "^3.11"
python-keycloak = "^3.7.0"
click = "^8.1.7"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
ruff = "^0.1.8"
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
